// app/(dashboard)/certificates/electrical-installation-condition-report/[id]/update/schedule-of-inspections/section-3/schema.ts
import { z } from "zod";

export const Schema = z.object({
  item_3_1A: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_1B: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_1C: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_1D: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_1E: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_1F: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_1G: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_1H: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_1I: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_2: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_3A: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_3B: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_3C: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_3D: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_3E: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_3_3F: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
});